% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gghazard.r
\name{fortify.cox.zph}
\alias{fortify.cox.zph}
\title{Create data structures from survival hazard Cox Regression}
\usage{
\method{fortify}{cox.zph}(x, df = 4, nsmo = 40)
}
\arguments{
\item{x}{result of the \code{cox.zph} function.}

\item{df}{the degrees of freedom for the fitted natural spline, \code{df=2}
leads to a linear fit.}

\item{nsmo}{number of points used to plot the fitted spline.}
}
\value{
\code{list} with residuals and fit+confidence intervals for each
        component of \code{x}
}
\description{
This function generates data to be used with \code{ggplot2} and is custom-
tailored for use with \code{gg_cox_zpn}. It returns a list the length of
\code{x} with names matching the row names of the output of \code{cox.zph}.\cr
\cr
Each \code{list} element has two \code{data.frames}, \code{resid} and
\code{pred} matching the residuals and fit+confidence interval for that element
along with \code{ylab} (which matches the list element name) and a boolean
\code{log} which indicates whether a log scale should be used.
}
\note{
Designed to be used with gg_cox_zph but can be used separately.
}
\examples{
library(survival)
library(ggplot2)
vfit <- coxph(Surv(time,status) ~ trt + factor(celltype) +
              karno + age, data=veteran, x=TRUE)

gg_cox_zph(fortify(cox.zph(vfit)))
}

