% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gghazard.r
\name{gg_cox_zph}
\alias{gg_cox_zph}
\title{Graphical Test of Proportional Hazards}
\usage{
gg_cox_zph(x, resid_col = "black", resid_shp = 1, pred_col = resid_col,
  pred_minmax_col = pred_Col, pred_minmax_linetype = "dashed",
  ribbon_col = pred_Col, ribbon_alpha = 0.25)
}
\arguments{
\item{x}{result of the \code{fortify.cox.zph} function.}

\item{resid_col}{color for the residuals points}

\item{resid_shp}{point shape for the residuals points}

\item{pred_col}{color for the prediction fit line}

\item{pred_minmax_col}{color for the confidence interval lines}

\item{ribbon_col}{color for the shaded confidence interval region}

\item{ribbon_alpha}{alpha value for the confidence interval region}

\item{pre_minmax_linetype}{line type for the confidence interval lines}
}
\value{
\code{grid} object
}
\description{
Displays a graph of the scaled Schoenfeld residuals, along with a smooth curve.
}
\examples{
library(survival)
library(ggplot2)
vfit <- coxph(Surv(time,status) ~ trt + factor(celltype) +
              karno + age, data=veteran, x=TRUE)

gg_cox_zph(fortify(cox.zph(vfit)))
}

