% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gghazard-package.r
\docType{package}
\name{gghazard}
\alias{gghazard}
\alias{gghazard-package}
\title{Improved Base and Grid Plots for Survival Hazard Cox Regression}
\description{
Improved Base and Grid Plots for Survival Hazard Cox Regression
}
\author{
Bob Rudis (@hrbrmstr)
}

