% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_cox_zph.r
\name{plot.cox.zph}
\alias{plot.cox.zph}
\title{Graphical Test of Proportional Hazards}
\usage{
\method{plot}{cox.zph}(x, resid = TRUE, se = TRUE, df = 4, nsmo = 40,
  var, smoothcol = "black", confcol = smoothcol, ...)
}
\arguments{
\item{x}{result of the \code{cox.zph} function.}

\item{resid}{a logical value, if \code{TRUE} the residuals are included on
the plot, as well as the smooth fit.}

\item{se}{a logical value, if \code{TRUE}, confidence bands at two standard
errors will be added.}

\item{df}{the degrees of freedom for the fitted natural spline, \code{df=2}
leads to a linear fit.}

\item{nsmo}{number of points used to plot the fitted spline.}

\item{var}{the set of variables for which plots are desired. By default,
plots are produced in turn for each variable of a model.
Selection of a single variable allows other features to be added to
the plot, e.g., a horizontal line at zero or a main title.\cr
\cr
This has been superseded by a subscripting method; see the example below.}

\item{smoothcol}{color for the smooth fit line}

\item{confcol}{color for the upper/lower confidence interval lines}

\item{...}{additional graphical arguments passed to the \code{plot} function.}
}
\description{
Displays a graph of the scaled Schoenfeld residuals, along with a smooth curve.
}
\details{
This is an augmented version of the \code{survival::plot.cox.zph} function
that allows for specifying separate colors for the smooth fit line and
confidence interval lines
}
\note{
Load this pacakge AFTER the \code{survival} pacakge.
}
\examples{
library(survival)
vfit <- coxph(Surv(time,status) ~ trt + factor(celltype) +
              karno + age, data=veteran, x=TRUE)

plot(cox.zph(vfit))
}

